feat: added visible_only option
